llvm_tools: Add llvm_local_bisection.sh
This adds a template for a complete bash script for llvm local
bisection.

This template is from a script that llozano gave to ryanbeltran,
but reworked to be easily extendable to other bisection steps.
Comes with "batteries included" build_llvm and "build_pkg"
functions.

BUG=None
TEST=Run on llvm-project as the `git bisect run` arg.

Change-Id: Ibc0115033d07b76d40ec2f4b7d7950cf9a518152
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3563907
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>