rust_uprev: Remove flip_mirror_in_ebuild
Previously, we temporarily switched on RESTRICT="mirror" for
an ebuild before updating the manifest. This allows us to fetch
distfiles from their original locations (rather than the default,
which is to require them to be fetched from mirrors). We don't
actually need this, so this change removes the code that does
this.

BUG=None
TEST=unit tests, also tested on rust-1.60.0 uprev

Change-Id: I5f29ffad83a5826dbe523db4657d9ea17c43bcff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3594132
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>