rust_uprev: Support links to base ebuild files
Previously, rust_uprev required that there be exactly one ebuild
file for a given rust{,-bootstrap,-uprev} version. This breaks
when using the common idiom of having a rust-1.2.3.ebuild and
a symlink to it with an -r version. This change adds support
for that by excluding such symlinks from the count, and removing
both the base files and the symlinks when removing ebuild versions.

BUG=b:227370760
TEST=Example uprev created by this script: crrev.com/c/3919575

Change-Id: I901ac8208c6bc48670860644ae91ee408e5f155d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3651375
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>