compiler_wrapper: add a --version_suffix flag to build.py
This allows us to provide more information about the compiler being
wrapped to the wrapper. The intended use of this is to embed hashes of
the compiler being wrapped into the wrapper itself, so as the compiler
changes, the wrapper's SHA changes with it.

While I'm in the area, fix up a comment, and apply this flag to our
manual wrapper installation script.

BUG=b:222321317
TEST=`emerge llvm`

Change-Id: I20380722e1e539d51fe7ea708c63dad696d84c87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3696434
Reviewed-by: Denis Nikitin <denik@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>