pgo_rust.py: add
This is a Python script to handle all aspects of creating and
benchmarking PGO profiles for the Rust compiler.

Note that this script is not usable without crrev.com/c/3762719, but as
no one else will need to use the script until the next Rust uprev, there
seems to be no need for Cq-Depend.

BUG=b:234046818
TEST=Used the script to enable PGO for Rust 1.62.1

Change-Id: I9a35cee3a2d6350a3a29b85c8d62ef450ce8fc22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3763629
Commit-Queue: Michael Benfield <mbenfield@google.com>
Tested-by: Michael Benfield <mbenfield@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>