llvm_tools: Remove unused llvm_patch_management.py
This code is not used by anything on the ChromeOS side,
and is just a wrapper around patch_manager.py itself.

The README implies that it auto-fills the command line
arguments, but this is somewhat unneccesary, and if we
want to add support for this workflow, it should likely
be in patch_manager.py itself as a subcommand.

This commit removes llvm_patch_management.py and
llvm_patch_management_unittest.py, and also
removes it from the README.md.

Also removes other incorrect flags from the README.md.

BUG=b:188465085
TEST=./run_tests_for.py llvm_tools/*

Change-Id: Ibf5f77977f70a8b7334e111a92a8fec5be462201
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3765939
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>