llvm_tools: Remove unused patch_metadata_file arg
This argument is not actually used anymore. Technically
it was never actually needed and it was passed in
other ways to the patch_manager.py even before the
refactoring. This just removes it from one function's
arguments.

It's hard to clean the forwarded patch_metadata_file arg
in every situation, so let's just do it for this function
that we introduced.

BUG=None
TEST=./run_tests_for llvm_tools/

Change-Id: I5c3f2a217df082cd51362aa70c8b585ad5135e47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3780680
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>