compiler_wrapper: redisable unused-but-set-var
This CL add the Wno-unused-but-set-variable flag back into the compiler
wrapper. It was previously removed, but there are still occurances of
the warning and so we were never able to bundle that change into
chromoumos-overlay/llvm. An attempt to revert the original CL failed due
to a merge conflict.

BUG=b:227655984
TEST=None

Change-Id: Ic148a6ae577d5e4394249693ce9b09b95cb9df16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3868978
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>