crosperf: find sshwatcher relative to chromeos_root
The root of the ChromiumOS source tree is not always in the same
location relative to where the crosperf scripts are. This makes
the logic for finding sshwatcher.go not work everywhere. To fix
this, this change uses the label.chromeos_root path that is
present in crosperf runs to find sshwatcher.go.

BUG=b:235119397
TEST=run crosperf in standalone toolchain-utils as well as
  inside ChromiumOS tree to see that sshwatcher.go is found
  in both cases.

Change-Id: I981a55d3cfd2638da0bf320b86ebc5601101ce3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3935652
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>