Revert "Revert "compiler-wrapper: adds an IWYU component""
This reverts commit 537f2ecfe7bc9eb996cd4abedae26bf699ab54dd.

It also fixes the logic bug which caused the reert in the first place.
IWYU flag prcessing was removing the file names if IWYU was not needed,
which meant clang tidy calls were not getting the file names added to
the back of the command. It also fixes the header to the most recent
style guidance and updates some golang variable names to make sure we
pass go lint.

BUG=b:237320348
TEST=go test

Change-Id: I14885fb90f97d2fb483ae1a01e14c44050852dc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/3997486
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>