compiler_wrapper: remove -fexperimental-new-pass-manager
We previously passed -fexperimental-new-pass-manager to Clang to turn on
the new pass manager. The new pass manager has been made the default,
and -fexperimental-new-pass-manager has become a no-op. In newer
versions of LLVM, it has been removed. This means these versions
of LLVM will fail when the flag is passed in. To make compiler_wrapper
compatible with these versions, remove the flag.

BUG=b:258854887
TEST=build new wrapper, build LLVM that doesn't have the flag, see it work

Change-Id: I2bb6de0075973b0b2cdb7700306b094ec1e08de6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4061509
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>