llvm_tools: Escape package category in log file
Currently, if using llvm_local_bisection.sh, if the
pkg in build_pkg contains a forward slash (like if using an atom
with the package category), it will error out complaining the
parent directory of the log file won't exist.

This isn't intended behaviour, and is a quick fix with a global
variable substring substitution.

BUG=None
TEST=Bisecting media-libs/${internal arc package}

Change-Id: Ifa9a7dccaf3091eebc6e0b975ec0f6f7d37a18c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4116246
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>