update_kernel_afdo: Move away branch dependent configurations
When we add a new kernel version to afdo pipeline we don't want
to update it in the branches since it's likely that afdo was enabled
in the same milestone.
In the current design configuration changes affected all branches and
to skip the update in branches we had to use "skip branches" tricks.

The better solution which is added here is to move away branch dependent
configurations from the script and read it from remote refs.

BUG=b:272295936
TEST=tested locally

Change-Id: I7856b15c50182f4897859022906e1263e76ec54e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4321430
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>