git_llvm_rev: Pass "--" to git to avoid "ambiguous argument" complaints
Some git comments can take revisions or paths in the same position, with
the same syntax. In such cases, if the command fails, git likes to complain
that the argument was ambiguous, and suggests disambiguating it using "--".
This is often not the most helpful error message - for example, if an
expected revision does not exist in the repository, it would be more
helpful to say that, rather than complain about ambiguity. To make
git's diagnostics more helpful, add "--" in a few commands.

BUG=None
TEST=Ran git_llvm_rev.py in a few cases to see it work.

Change-Id: I06a457b74319fec16ddffc72f64679bd39a282e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4386394
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>