llvm_tools: Add manifest_utils
This is part of the effort to get LLVM 9999 ebuilds working.
Because LLVM 9999 ebuilds need to know what revision LLVM is at,
and because that info is only available during unpack time currently,
we need to move the LLVM_HASH into the manifest files.

Having the LLVM_HASH in the manifest allows repo init to pull down
the correct LLVM revision in the source tree checkout. This then
allows chromeos-version.sh to determine what the new ebuild name
should be during the PUpr rename stage. We could either parse
the ebuild via chromeos-version.sh, or we could update the manifest.

This is part of the effort for the latter implementation.

This library contains functions that let us locate and update
the manifest file, which we want to be as seamless as possible
for toolchain maintainers in ChromeOS.

BUG=b:275877115
TEST=manifest_utils_unittest.py

Change-Id: I45b8ebc72d5e6462552e6873d897199481007f46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4575384
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>