llvm_tools: Update repo manifest with llvm hash
At present, the only parts of the ChromeOS tree that
update_chromeos_manifest needs to edit are the LLVM_HASH
and LLVM_NEXT_HASH entries. However, as we transition to LLVM-9999
ebuilds, we will need to also update the repo manifest files
to the same hash for the PUpr builder.

This introduces a new flag for update_chromeos_llvm_hash, called
--repo_manifest, which sets the llvm-project revision to
whatever is set as the LLVM git sha. Then it uploads this
change for review, alongside the LLVM_*_HASH changes.

BUG=b:275877115
TEST=./update_chromeos_llvm_hash.py \
     --is_llvm_next --llvm_version google3 \
     --failure_mode continue --repo_manifest
TEST=update_chromeos_llvm_hash_unittest.py

Change-Id: Iebdd020e39e40f6241d4a02e2177233b00373ba5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4575386
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>