cros_utils: Make GetLogger have type guarantees
At present, GetLogger could return None, Logger, or MockLogger. No code
seems to be built for when GetLogger returns None, and it would instead
be nice to have that safety.

The original InitLogger relies on modifying global state to ensure
the main_logger is not None. This does not make the type analysis happy,
so instead lets have InitLogger return the logger we initted.

This also does a minor clean up of the code.

BUG=None
TEST=./run_tests_for.py cros_utils/*

Change-Id: I2f5b142e088ceb0299d7ac2513b7c3c6f3800800
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4674780
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>