rust_uprev: fetch Rust sources from upstream if not present on local mirror
Previously, fetching sources from upstream, verifying their integrity,
and uploading them to the local mirror was a manual process. With this
change, this happens automatically. The integrity of the sources is
verified using GPG, which checks that the sources have the expected
checksum, that the checksum information was signed by the signing
key used by the Rust project, and that the key and signature are
current (that is, the key has not been revoked and the key and
signature have not expired). The expected key is hardcoded in the
script so that if we ever encounter sources signed by a different
key, the verification will fail and we will have to manually
verify that the new key is legitimate.

BUG=b:271016462
TEST=unittest, use new script to create rust 1.71.1 uprev

Change-Id: I9b2129ed82ca7de9f9aadfd275683f49cb72561a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4767736
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>