pgo_rust: fix line-too-long pylint complaints
A number of lines in pgo_rust triggered line-too-long complaints from
pylint. This fixes those, with the exception of one URL which happens
to be longer than the line length. Shortening the URL is possible,
but would make it less clear, so I've opted to instead silence the
complaint for that line.

BUG=None
TEST=repo upload # see that cros lint no longer complains

Change-Id: Ia6433eae378f66376aa33031cb3cee66e42bb69c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4775778
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>