rust_uprev.py: run outside the chroot
This modifies rust_uprev.py to run outside the chroot instead of inside
the chroot. This change is necessary, because repo is no longer
available inside the chroot.

This also refactors the code a little bit: template_version is now
determined before prepare_uprev() instead of inside
prepare_uprev(). This is done so that build_cross_compiler() does not do
its own separate determination of which rust ebuild to process.

BUG=b:296186356
TEST=use with --restart and --continue to verify it works

Change-Id: I015140f745a3a4dff2d1ead46142c8fd090ae29c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4786827
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>