Update check_portable_toolchains in python
This script originally was a shell script in crostc/contrib.
This is now a useful smoketest to see if the SDK tarballs
for simple chrome actually work. Formalise this, rewrite
in python, and update the paths for the new SDKBuilder pipeline.

Originally this script was called check_sdk_toolchains,
however this name didn't really describe what it actually did.

BUG=b:299317450
TEST=./check_portable_toolchains '2023.09.01.221258'

Change-Id: I5fa3db977011ca973b17a82fdbf83b881b2eb5f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/4847972
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>