rust_tools: add automatic rust-bootstrap updating
This functionality will search for newly-committed versions of
dev-lang/rust, and add corresponding rust-bootstrap ebuilds.

I originally thought it'd be good to search gs:// for src tarballs, but
there's no pressing need to do that over waiting for rust versions to be
landed in-tree. Waiting for new rust versions in-tree also means we
don't have to worry as much about:
- signature verification on tarballs (dev-lang/rust landing means
  everything looks good)
- rust-bootstrap tarballs not WAI in CrOS (if it works as dev-lang/rust,
  it works for the SDK)

BUG=b:309950482
TEST=Ran the script, unittests

Change-Id: I88d66f2963f348428059c891e11d0d2a2f075e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5105284
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>