contrib: add bgtask
Adding this at the request of ajordanr@ on the 'improving the Mage
experience' doc.

We didn't have a contrib folder here, so I added one. We won't be able
to bypass CR+2, but that's probably fine.

I also sorted .gitignore while I'm in the area.

BUG=b:315871489
TEST=Ran the program.

Change-Id: I77146550cfaf9879d38ac15b4b308a4a4aabb897
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5117045
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>