llvm_tools: add setup_for_workon.py
This adds a script to complement the changes made in
crrev.com/c/5120358. The intent is for devs to run this script prior to
hacking on llvm.

There are no tests in this CL since this script is _mostly_ just a
series of `subprocess.run`s with a tiny bit of intervening logic. It's
also 100% intended for interactive use.

BUG=b:315871489
TEST=Ran the script.

Change-Id: I2d516b22ba7f7d95d3a3666cc2b6ab51810cf7ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5119366
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>