llvm_tools: add llvm_simple_bisect.py
Normally for bisecting llvm we use llvm_local_bisection script with
modifications for the steps we are running and the tests we want to
perform, which works fine but requires some setup and it's easy to make
mistakes.

This tool may be a little less versatile than that one, but it
simplifies the majority of usecases including:
* Bisecting a command such as "emerge my_package" that fails with a
  specific message for some versions of LLVM
* Bisecting a failure in building LLVM itself

The test command can be replaced with a script if desired to test
complicated things such as correctness, build times, etc.

What's not supported by this tool as of yet and therefore justifies
having the old one as well still is full customization of the llvm build
command, including hooks that run before building llvm. Some aspects of
the LLVM build are configurable including certain USE flags.

BUG=b:316209470
TEST=Used to bisect a failure in chromeos-base/ml

Change-Id: I93b51cff186bcc2379b1a73b98a5d518ef4f3f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5120397
Auto-Submit: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>