llvm_tools: add a -Werror log parsing tool
This is the first "half" of points 3 and 4 in the description of
b/316172255. This is useful on its own, since if a Mage runs a local
`FORCE_DISABLE_WERROR` build, they can point this tool at their build
tree and get a full report of all of the suppressed warnings.

BUG=b:316172255
TEST=Unittests; ran on a few `FORCE_DISABLE_WERROR` trees

Change-Id: I95bfc5de70f8614bf0ae7570fd5ab31766ce405d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5180185
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>