rust_uprev: refactor repetitive routines regarding rusty revisions
rust_uprev handles several ebuilds: dev-lang/rust,
dev-lang/rust-artifacts, dev-lang/rust-host, and
dev-lang/rust-bootstrap. Excepting the last, the script generally
needs to process all of them the same way. Previously, this was done
by repeating code snippets for each package. This change instead
puts designators for the packages to process in a list, then
iterates over that list and runs the same code on each. This makes
it easier to ensure that all packages are indeed processed the same
way, instead of, say, adding code that creates a new ebuild for a
package, but not adding code that removes the old version.

BUG=b:322348600
TEST=presubmit, create a rust uprev with it

Change-Id: Ide5b2dab00e88633c8e5b80c1ee465adee14917b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5255518
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>