rust_uprev: remove update_manifest()
update_manifest() takes a path, which we then translate to a package
name, which we then look up in the filesystem to get the path to
the ebuild file. This change removes a level of indirection by
calling ebuild_actions() with the package name directly.

BUG=None
TEST=Create a rust uprev

Change-Id: I1b1651cc7d1fc735c353b55fa28c5766015e4829
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5255377
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>