pgo_rust: explicitly build rust-artifacts
pgo_rust builds dev-lang/rust and dev-lang/rust-host a couple of
times with different USE flags (instrumented for frontend profiling,
instrumented for LLVM profiling) and generates profile data with each
configuration.  However, USE flags are not taken into account when
deciding if dependencies need to be rebuilt. This resulted in
dev-lang/rust-artifacts not being rebuilt with the changed USE flags,
and profiles not being generated correctly. This change explicitly
rebuilds dev-lang/rust-artifacts each time we change the flags, so
that profiles are generated correctly.

BUG=b:322998657
TEST=pgo_rust.py generate; see that profiles are generated successfully

Change-Id: I28265c8d5cf98c4abf751f681e7a361be587359a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5258978
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>