rust_uprev: don't update rust-bootstrap
rust_uprev.py contains code that updates the rust-bootstrap ebuild.
This was previously necessary, because, without it, the version of
rust-bootstrap needed to build the other rust packages would not
be available. However, this is no longer true: We now update
rust-bootstrap to the same version as dev-lang/rust and the others,
which means that when we advance the non-bootstrap version, it
will only be one step ahead of rust-bootstrap and therefore work.
auto_update_rust_bootstrap.py will then take care of updating
rust-bootstrap. As a result, we don't need logic for that in
rust_uprev.py anymore, and this change removes it. Some code that
refers to rust-bootstrap is still preserved, namely the code
updates the bootstrap version encoded in cros-rustc.eclass and
the code that verifies that rust-bootstrap sources and prebuilts
are available on the mirror.

BUG=None
TEST=presubmit

Change-Id: I0959bfc6d442467622b0403e6cb975d5b579aad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5259526
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>