llvm_tools: add package stabilization script
This allows a dev to make all changes in their LLVM ebuilds stable with
a single command. Critically, if used with `--llvm-next`, a user can run
`cros-tree bootstrap-sdk` in the tree this was run in, and they'll get
something very close to an SDK run with llvm-next, but locally.

No tests are included, since this has near zero conditional logic. Vast
majority of that is elsewhere.

BUG=b:333737743
TEST=Ran the script with --llvm-next, with an updated llvm-next.
TEST=New ebuilds were created, as expected.

Change-Id: Iad7bdcd4c51828e185fea698cd43922be47dd75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5444011
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Ryan Beltran <ryanbeltran@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>