rust-analyzer-chromiumos-wrapper: Perform exact prefix match for replacement
Paths in rust-analyzer RPC messages can be either file: URIs or normal
paths. Previously both were handled by allowing replacements in the
middle of a string; but this was an ugly solution as it was explained in
the comments.

Hence, introduce a separate path to process file: URIs, and perform
exact prefix match for both URIs and paths.

The assumption that a path can be a URI or a normal path, and not
something else (e.g. JSON object encoded in string that contains
multiple paths), was validated through instrumenting the previous code
such that it will error out on such unexpected entries. The instrumented
version worked in VS Code rust-analyzer without hitting the exception
path.

BUG=b:333979840
TEST=cargo test
TEST=check ctrl+click navigation in VS Code in CrOS repository

Change-Id: I556def4e0a0ea9d7333a2bfb30f0bc481bc20cc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5449292
Tested-by: Tatsuyuki Ishi <ishitatsuyuki@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Auto-Submit: Tatsuyuki Ishi <ishitatsuyuki@google.com>