remove image_chromeos
The only critical feature I see here that isn't covered by the standard
`cros flash` command is related to machine locking, which is no longer
something that we (the toolchain team) need to care about.

If we want a script for this, it seems there's tons of room to simplify
this script, at least.

BUG=b:337837895
TEST=None

Change-Id: I6b31e3adae089f011dfd03eefff3261dcc6b46d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5496645
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>