pgo_tools: allow auto_update script to force PGO generation
At the moment, we generate probably 4 LLVM PGO profiles per year. This
leaves a lot of room for things to silently break the PGO script
(b/335298562 & b/335515781). It seems worthwhile to have a cronjob
forcibly run PGO generation once a week or so, even if its result will
be discarded, so we can have confidence that the pipeline is
continuously working well.

BUG=b:334876457
TEST=Ran the script with llvm-next pointed at the current LLVM version,
TEST=with the new flag.

Change-Id: I96c437e314b43f8398602beab871eaffff5742bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5502775
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>