check-presubmit: install mypy earlier; remove deps
`mypy` is used more often now; it should probably be installed with
everything else. It also just so happens that "everything else" is the
empty set, so this can be cleaned up a bit.

This also fixes a latent bug with `--install_deps_only`: mypy wasn't
installed by this.

BUG=None
TEST=repo upload

Change-Id: I109d1d914b8b485613f4e3c4da20fbd6a474d0e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5507514
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>