push `FindChromeOSRootAbove` out of hash-getting
Most scripts have a clear idea of what the CrOS root directory is. Pass
that in here, rather than taking anything in a CrOS tree, so this is
easier to think about.

I hand-audited the handful of calls to this function. Most were already
doing this.

BUG=None
TEST=Unittests

Change-Id: Id0dd6c13bf3c78c828594fefd6a5f523f9553333
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5533680
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>