llvm_tools: Fix writing to cherry/ subdirs
The parent of each atomic_write call must exist prior to calling
the atomic_write call. When testing, I found that this broke
when the cherry/ patch subdirectory didn't exist.

Not likely a problem in real-world usage, but it's a nice fix
for testing.

BUG=b:343568613
TEST=py/bin/llvm_tools/create_patch_file.py \
     --patch-dir patches_test

Change-Id: I060c3c98c9651c6bbec7dde7150aeb076f6edab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5636515
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>