llvm_tools: Add ready_llvm_branch.py
This script creates an LLVM branch locally with the necessary patch
stack.

This is a really simple script that's basically just a patch_utils.py
wrapper--and is almost entirely a subset of patch_manager.py. But it
has more ergonomic CLI usage, so I thought it would be still nice to
add.

BUG=b:343568613
TEST=py/bin/llvm_tools/ready_llvm_branch.py \
     -r 516547 --branch-number 2

Change-Id: I610b3989da47de004a5c9d8491ad6dfbb04c3ef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5685179
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>