generate_pgo_profile: update absl/gtest
absl breaks with the newer libcxx version (likely this wasn't identified
because the bootstrap SDK was just updated to llvm-next a few days ago,
and this script is run in a bootstrap SDK). Pick a newer version for
that.

Further, the newer absl version uses newer gtest features. Bump that to
a version from 2024, rather than 2021.

BUG=b:366159701
TEST=generate_pgo_profile.py

Change-Id: I54d94348bdfdffc7f97d4644dd7eb08634baead1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5858177
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>