pgo_tools: replace --rev with --branch in pgo generation script
This script unconditionally tries to apply patches to a SHA. This patch
application logic is broken (and needs to be fixed), but at the same
time, "check out this branch and don't apply patches" needs to be
implemented.

Both could be supported simultaneously (which was my original plan), but
I can't see a clear use for keeping `--rev` around. Happy to readd
`--rev` as an alternative to `--branch` if a use-case appears.

BUG=b:372627719
TEST=Unittests

Change-Id: I4e7124f13b55dd96bb58c6f044f9859e28156eaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/5931005
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>