githooks: force py3.8
py3.11 doesn't work for this script yet, due to lack of ensurepip.

This can't be done in the `#!` line, as this script needs to be runnable
outside of the chroot. This workaround isn't 100% (if there's no reexec
inside of the chroot, we don't have a well-specified Python version),
but should unbreak the CQ

BUG=b:378723931
TEST=Ran presubmits

Change-Id: I4da1136de72df072f5b53ffa0942a6e3fc83cf77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6011894
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>